+++ pkg/cmd/delete/clusterclaim/cmd.go
@@ -1,38 +1,48 @@
 // Copyright Contributors to the Open Cluster Management project
-package clusterpoolhosts
+package clusterclaim
 
 import (
 	"fmt"
+	"path/filepath"
+
+	clusteradmhelpers "open-cluster-management.io/clusteradm/pkg/helpers"
 
 	"github.com/open-cluster-management/cm-cli/pkg/clusterpoolhost"
 	genericclioptionscm "github.com/open-cluster-management/cm-cli/pkg/genericclioptions"
 	"github.com/open-cluster-management/cm-cli/pkg/helpers"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
 
+	"github.com/open-cluster-management/cm-cli/pkg/cmd/create/cluster/scenario"
+
 	"github.com/spf13/cobra"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 )
 
-var example = `
-# Get cluster pool hosts
-%[1]s get clusterpoolhosts
-`
-
 const (
-	scenarioDirectory = "clusterpoolhost"
+	scenarioDirectory = "create"
 )
 
-// NewCmd provides a cobra command wrapping NewCmdImportCluster
+var valuesTemplatePath = filepath.Join(scenarioDirectory, "values-template.yaml")
+
+var example = `
+# Delete a clusterclaim in the current clusterpoolhost
+%[1]s delete clusterclaim|cc <clusterclaim_name>[,<clusterclaim_name>...] <options>
+
+# Delete a clusterclaim on a given clusterpoolhost
+%[1]s create clusterclaim|cc <clusterclaim_name>[,<clusterclaim_name>...] <clusterpoolhost_name> <options>
+`
+
+// NewCmd ...
 func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOStreams) *cobra.Command {
 	o := newOptions(cmFlags, streams)
-
 	cmd := &cobra.Command{
-		Use:          "clusterpoolhosts",
-		Aliases:      []string{"cphs"},
-		Short:        "list the clusterpoolhosts",
+		Use:          "clusterclaim",
+		Aliases:      []string{"cc", "clusterclaims"},
+		Short:        "Delete clusterclaims",
 		Example:      fmt.Sprintf(example, helpers.GetExampleHeader()),
 		SilenceUsage: true,
 		PreRunE: func(cmd *cobra.Command, args []string) error {
+			clusteradmhelpers.DryRunMessage(cmFlags.DryRun)
 			return clusterpoolhost.BackupCurrentContexts()
 		},
 		Run: func(cmd *cobra.Command, args []string) {
@@ -45,6 +55,7 @@ func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOSt
 		},
 	}
 
-	cmd.Flags().BoolVar(&o.raw, "raw", false, "If set return a raw display")
+	cmd.SetUsageTemplate(clusteradmhelpers.UsageTempate(cmd, scenario.GetScenarioResourcesReader(), valuesTemplatePath))
+
 	return cmd
 }