+++ pkg/cmd/delete/clusterclaim/exec.go
@@ -1,0 +1,58 @@
+// Copyright Contributors to the Open Cluster Management project
+package clusterclaim
+
+import (
+	"fmt"
+
+	"github.com/open-cluster-management/cm-cli/pkg/clusterpoolhost"
+
+	"github.com/spf13/cobra"
+)
+
+func (o *Options) complete(cmd *cobra.Command, args []string) (err error) {
+	if len(args) < 1 {
+		return fmt.Errorf("clusterclaim name is missing")
+	}
+	o.ClusterClaims = args[0]
+	if len(args) > 1 {
+		o.ClusterPoolHost = args[1]
+	}
+	return nil
+}
+
+func (o *Options) validate() error {
+	return nil
+}
+
+func (o *Options) run() (err error) {
+	cphs, err := clusterpoolhost.GetClusterPoolHosts()
+	if err != nil {
+		return err
+	}
+
+	currentCph, err := cphs.GetCurrentClusterPoolHost()
+	if err != nil {
+		return err
+	}
+
+	if len(o.ClusterPoolHost) != 0 {
+		cph, err := cphs.GetClusterPoolHost(o.ClusterPoolHost)
+		if err != nil {
+			return err
+		}
+
+		err = cphs.SetActive(cph)
+		if err != nil {
+			return err
+		}
+	}
+	err = clusterpoolhost.DeleteClusterClaims(o.ClusterClaims, o.CMFlags.DryRun, o.outputFile)
+	if err != nil {
+		return err
+	}
+
+	if len(o.ClusterPoolHost) != 0 {
+		return cphs.SetActive(currentCph)
+	}
+	return nil
+}